fix: Do not leak global declared types into distribution bundle
